UnionWrappers

Documentation for UnionWrappers.

UnionWrappers.AbstractEltypeWrapper

UnionWrappers.AbstractSizeWrapper

UnionWrappers.AbstractUnionWrapper

UnionWrappers.unwrap

UnionWrappers.wrap_eltype

UnionWrappers.wrap_size

UnionWrappers.wrap_union

UnionWrappers.wrapped_union

UnionWrappers.AbstractEltypeWrapper — Type

AbstractEltypeWrapper{E,U} <: AbstractUnionWrapper{U}

Wrapper that stores an additional element type as a type parameter. The element type can be queried using eltype(w). There are implementations of wrap_eltype for NTuple and NamedTuple The abbreviation alias EWrap{U} can shorten function signatures.

UnionWrappers.AbstractSizeWrapper — Type

AbstractSizeWrapper{ND,D,E,U} <: AbstractEltypeWrapper{E,U}

Wrapper that stores an additional number of dimensions, ND == length(D), and sizes of dimensions, D, as type parameters. The size can be queried by size(w) and the number of elements = prod(D) can be queried using length(w). Although ND is redundant to D, it is stored to allow dispatch on number of dimensions. There are implementations of wrap_size for ComponentArray.

UnionWrappers.AbstractUnionWrapper — Type

AbstractUnionWrapper{U}

Basic Wrapper that stores a single type parameter to distinguish different wrapped types for dispatch. The default wrap_union creates a wrapper for uniontype Any. The abbreviation alias UWrap{U} can shorten function signatures.

UnionWrappers.unwrap — Method

wrap_union(x)

Wraps into a AbstractUnionWrapper type to avoid compilation for types with much information in type parameters.

Specific methods create different Wrappers with a small amount of information on the type and its properties preserved in type parameters.

The default produces as UnionWrapper{Val{:Any}()}.

See unwrap(w) and wrapped_union(w) to extract the original value or the uniontype of the wrapped object.

UnionWrappers.wrap_eltype — Function

AbstractEltypeWrapper{E,U} <: AbstractUnionWrapper{U}

Wrapper that stores an additional element type as a type parameter. The element type can be queried using eltype(w). There are implementations of wrap_eltype for NTuple and NamedTuple The abbreviation alias EWrap{U} can shorten function signatures.

UnionWrappers.wrap_size — Function

AbstractSizeWrapper{ND,D,E,U} <: AbstractEltypeWrapper{E,U}

Wrapper that stores an additional number of dimensions, ND == length(D), and sizes of dimensions, D, as type parameters. The size can be queried by size(w) and the number of elements = prod(D) can be queried using length(w). Although ND is redundant to D, it is stored to allow dispatch on number of dimensions. There are implementations of wrap_size for ComponentArray.

UnionWrappers.wrap_union — Function

wrap_union(x)

Wraps into a AbstractUnionWrapper type to avoid compilation for types with much information in type parameters.

Specific methods create different Wrappers with a small amount of information on the type and its properties preserved in type parameters.

The default produces as UnionWrapper{Val{:Any}()}.

See unwrap(w) and wrapped_union(w) to extract the original value or the uniontype of the wrapped object.

UnionWrappers.wrapped_union — Method

wrap_union(x)

Wraps into a AbstractUnionWrapper type to avoid compilation for types with much information in type parameters.

Specific methods create different Wrappers with a small amount of information on the type and its properties preserved in type parameters.

The default produces as UnionWrapper{Val{:Any}()}.

See unwrap(w) and wrapped_union(w) to extract the original value or the uniontype of the wrapped object.