UnionWrappers

Documentation for UnionWrappers.

UnionWrappers.AbstractEltypeWrapper

UnionWrappers.AbstractLengthWrapper

UnionWrappers.AbstractUnionWrapper

UnionWrappers.AnyWrapper

UnionWrappers.ComponentArrayWrapper

UnionWrappers.ComponentVectorWrapper

UnionWrappers.NTupleWrapper

UnionWrappers.NamedTupleWrapper

UnionWrappers.unwrap

UnionWrappers.wrap

UnionWrappers.wrapped_type

UnionWrappers.AbstractEltypeWrapper — Type

AbstractEltypeWrapper{E,T} <: AbstractUnionWrapper{T}

Wrapper that stores an additional element type as a type parameter. The element type can be queried using eltype(w). For instance, see ComponentArrayWrapper.

UnionWrappers.AbstractLengthWrapper — Type

AbstractLengthWrapper{N,E,T} <: AbstractEltypeWrapper{E,T}

Wrapper that stores an additional number of elements as a type parameter. The number of elements can be queried using length(w). For instance, see NTupleWrapper.

UnionWrappers.AbstractUnionWrapper — Type

AbstractUnionWrapper{T}

Basic Wrapper that stores a single type parameter to distinguish different wrapped types for dispatch. For instance, see AnyWrapper.

UnionWrappers.AnyWrapper — Type

AnyWrapper = UnionWrapper{Val(:Any)}

Alias for the UnionWrapper that is created by wrap without a more specific method.

UnionWrappers.ComponentArrayWrapper — Type

ComponentArrayWrapper{E} = EltypeWrapper{E,Val(:ComponentArray)}

Alias for the EltypeWrapper that is created by wrap on a ComponentArray. For a ComponentVector the more specific type ComponentVectorWrapper{N,E} is created.

UnionWrappers.ComponentVectorWrapper — Type

ComponentVectorWrapper{N,E} = LengthWrapper{N,E,Val(:ComponentVector)}

Alias for the LengthWrapper that is created by wrap for a ComponentVector.

UnionWrappers.NTupleWrapper — Type

NTupleWrapper{N,E} = LengthWrapper{N,E,Val(:NTuple)}

Alias for the LengthWrapper that is created by wrap for an NTuple.

UnionWrappers.NamedTupleWrapper — Type

NamedTupleWrapper{N,E} = LengthWrapper{N,E,Val(:NamedTuple)}

Alias for the LengthWrapper that is created by wrap for a NamedTuple{K,NTuple{N,E}}.

UnionWrappers.unwrap — Method

wrap(x)

Wraps into a AbstractUnionWrapper type to avoid compilation for types with much information in type parameters.

Specific methods create different Wrappers with a small amount of information on the type and its properties preserved in type parameters.

The default produces as UnionWrapper{Val{:Any}()}.

See unwrap(w) and wrapped_type(w) to extract the original value or its type.

UnionWrappers.wrap — Function

wrap(x)

Wraps into a AbstractUnionWrapper type to avoid compilation for types with much information in type parameters.

Specific methods create different Wrappers with a small amount of information on the type and its properties preserved in type parameters.

The default produces as UnionWrapper{Val{:Any}()}.

See unwrap(w) and wrapped_type(w) to extract the original value or its type.

UnionWrappers.wrapped_type — Method

wrap(x)

Wraps into a AbstractUnionWrapper type to avoid compilation for types with much information in type parameters.

Specific methods create different Wrappers with a small amount of information on the type and its properties preserved in type parameters.

The default produces as UnionWrapper{Val{:Any}()}.

See unwrap(w) and wrapped_type(w) to extract the original value or its type.